api-m_s/requirement.txt
certifi==2019.11.28
chardet==3.0.4
Django==2.2.8
django-tastypie==0.14.2
djangorestframework==3.10.3
gunicorn==20.0.4
httpie==1.0.3
idna==2.8
Pygments==2.5.2
python-dateutil==2.8.1
python-mimeparse==1.6.0
pytz==2019.3
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7
